refactor(constand): 전체 로또와 관련된 숫자들 상수화
로또 최소 수, 최대 수, 로또 사이즈, 가격, 상금에 대한 숫자들을 상수화 하였습니다.

+++ src/main/java/lotto/domain/rule/result/RankResult.java
@@ -23,10 +23,10 @@ public record RankResult(Map<Rank, Integer> rankStatistics) {
     public int getTotalPrize() {
         int total = 0;
         for (Map.Entry<Rank, Integer> entry : rankStatistics.entrySet()) {
-            Long prize = entry.getKey().getPrize();
+            int prize = entry.getKey().getPrize();
             int value = entry.getValue();
 
-            total += (int) (prize * value);
+            total += (prize * value);
         }
         return total;
     }
